Remove Managing Processes section.
In general, SMS does not prescribe certain libraries or enforce the
use of certain gems or libraries in favor of developer
exploration and autonomy.

Developers should be encouraged to learn about the ecosystem
and should also be responsible enough to find or take ownership
of well maintained gems and use them to solve a specific problem
for the client while maximizing the value provided to the client.

Signed-off-by: Joseph Jaber <mail@josephjaber.com>