Add item to avoid more than 1 level of nesting in Routes.
It has been recommended by members of the Rails core team
not to use more than one level of route nesting. This practice lends itself to an opaque workflow and unnecessary bugs.

A resource will not need to be nested more than 1 level, as a
shadow hierarchy will provide the same utility. Resources that are
deeply nested may also be indicative of having unRESTful
resources.

Signed-off-by: Joseph Jaber <mail@josephjaber.com>