Remove item on discouraging overriding of ActiveRecord default.
A precedent has already been set in SMS code and culture to not
override ActiveRecord defaults. SMS Developers also tend not to
do so unless they are forced to, based on the context of the current
use case.

This recommendation might make the style guide less engaging
and may even bore many SMS developers.

Signed-off-by: Joseph Jaber <mail@josephjaber.com>