Add information on how to structure and organize classes.
Class composition is ordered based on how unlikely, dangerous, or
anomalous a particular piece of a class might be.

For example, including a module or utilizing a 3rd party gem macro
is less likely to be encountered and more likely to cause developer
confusion during the debugging process vs adding a
`belongs_to` association, which is both common and expected in
a Rails applications. For these reasons the former would be higher
in the class definition than the latter.

Signed-off-by: Joseph Jaber <mail@josephjaber.com>