Remove recommendation of foreign key constraint gem.
The Ruby gem ecosystem changes rapidly. Do not prescribe the
usage of a certain gem in favor of developer creativity and
autonomy to evaluate and choose a gem that best fits their use
case.

Database-specific gems make a project less database agnostic and
more difficult to work on, especially in the case of database
that SMS does not have license to use on development machines.

Signed-off-by: Joseph Jaber <mail@josephjaber.com>